Task 101863: Criar endpoint para obter cargos e funções
