+++ SME.ConectaFormacao.Aplicacao/Consultas/Proposta/ObterCargoFuncao/ObterCargoFuncaoQuery.cs
@@ -1,0 +1,16 @@
+﻿using MediatR;
+using SME.ConectaFormacao.Aplicacao.Dtos.Proposta;
+using SME.ConectaFormacao.Dominio.Enumerados;
+
+namespace SME.ConectaFormacao.Aplicacao
+{
+    public class ObterCargoFuncaoQuery : IRequest<IEnumerable<CargoFuncaoDTO>>
+    {
+        public ObterCargoFuncaoQuery(CargoFuncaoTipo? tipo)
+        {
+            Tipo = tipo;
+        }
+
+        public CargoFuncaoTipo? Tipo { get; }
+    }
+}
