+++ SME.ConectaFormacao.Aplicacao/Interfaces/Proposta/ICasoDeUsoObterCargoFuncao.cs
@@ -1,0 +1,10 @@
+﻿using SME.ConectaFormacao.Aplicacao.Dtos.Proposta;
+using SME.ConectaFormacao.Dominio.Enumerados;
+
+namespace SME.ConectaFormacao.Aplicacao.Interfaces.Proposta
+{
+    public interface ICasoDeUsoObterCargoFuncao
+    {
+        Task<IEnumerable<CargoFuncaoDTO>> Executar(CargoFuncaoTipo? tipo);
+    }
+}
