+++ SME.ConectaFormacao.TesteIntegracao/CasosDeUso/Usuario/ServicosFakes/AlterarEmailServicoAcessosCommandHandlerFake.cs
@@ -1,0 +1,13 @@
+﻿using MediatR;
+using SME.ConectaFormacao.Aplicacao;
+
+namespace SME.ConectaFormacao.TesteIntegracao.CasosDeUso.Usuario.ServicosFakes
+{
+    public class AlterarEmailServicoAcessosCommandHandlerFake : IRequestHandler<AlterarEmailServicoAcessosCommand, bool>
+    {
+        public Task<bool> Handle(AlterarEmailServicoAcessosCommand request, CancellationToken cancellationToken)
+        {
+            return Task.FromResult(true);
+        }
+    }
+}
