+++ SME.ConectaFormacao.TesteIntegracao/CasosDeUso/Usuario/ServicosFakes/ObterMeusDadosServicoAcessosPorLoginQueryHandlerFake.cs
@@ -1,0 +1,18 @@
+﻿using MediatR;
+using SME.ConectaFormacao.Aplicacao;
+using SME.ConectaFormacao.Aplicacao.DTOS;
+using SME.ConectaFormacao.TesteIntegracao.CasosDeUso.Usuario.Mocks;
+
+namespace SME.ConectaFormacao.TesteIntegracao.CasosDeUso.Usuario.ServicosFakes
+{
+    public class ObterMeusDadosServicoAcessosPorLoginQueryHandlerFake : IRequestHandler<ObterMeusDadosServicoAcessosPorLoginQuery, DadosUsuarioDTO>
+    {
+        public Task<DadosUsuarioDTO> Handle(ObterMeusDadosServicoAcessosPorLoginQuery request, CancellationToken cancellationToken)
+        {
+            if (request.Login == UsuarioMeusDadosMock.Login)
+                return Task.FromResult(UsuarioMeusDadosMock.DadosUsuarioDTO);
+
+            return Task.FromResult(new DadosUsuarioDTO());
+        }
+    }
+}
