--- SME.ConectaFormacao.TesteIntegracao/Usuario/Mocks/UsuarioMeusDadosMock.cs
@@ -1,32 +1,0 @@
-﻿using Bogus;
-using Bogus.Extensions.Brazil;
-using SME.ConectaFormacao.Aplicacao.DTOS;
-
-namespace SME.ConectaFormacao.TesteIntegracao.Usuario.Mocks
-{
-    public class UsuarioMeusDadosMock
-    {
-        public static string Login { get; set; }
-
-        public static string LoginNaoEncontrado { get; set; }
-
-        public static DadosUsuarioDTO DadosUsuarioDTO { get; set; }
-
-        public static void Montar()
-        {
-            var pessoa = new Person("pt_BR");
-
-            Login = pessoa.FirstName;
-
-            DadosUsuarioDTO = new DadosUsuarioDTO
-            {
-                Nome = pessoa.FullName,
-                Login = pessoa.FirstName,
-                Email = pessoa.Email,
-                Cpf = pessoa.Cpf()
-            };
-
-            LoginNaoEncontrado = new Person("pt_BR").FirstName;
-        }
-    }
-}
