--- SME.ConectaFormacao.TesteIntegracao/Usuario/ServicosFakes/AlterarEmailServicoAcessosCommandHandlerFake.cs
@@ -1,13 +1,0 @@
-﻿using MediatR;
-using SME.ConectaFormacao.Aplicacao;
-
-namespace SME.ConectaFormacao.TesteIntegracao.Usuario.ServicosFakes
-{
-    public class AlterarEmailServicoAcessosCommandHandlerFake : IRequestHandler<AlterarEmailServicoAcessosCommand, bool>
-    {
-        public Task<bool> Handle(AlterarEmailServicoAcessosCommand request, CancellationToken cancellationToken)
-        {
-            return Task.FromResult(true);
-        }
-    }
-}
