--- SME.ConectaFormacao.TesteIntegracao/Usuario/ServicosFakes/AlterarSenhaServicoAcessosCommandHandlerFaker.cs
@@ -1,19 +1,0 @@
-﻿using MediatR;
-using SME.ConectaFormacao.Aplicacao;
-using SME.ConectaFormacao.TesteIntegracao.Usuario.Mocks;
-
-namespace SME.ConectaFormacao.TesteIntegracao.Usuario.ServicosFakes
-{
-    internal class AlterarSenhaServicoAcessosCommandHandlerFaker : IRequestHandler<AlterarSenhaServicoAcessosCommand, bool>
-    {
-        public Task<bool> Handle(AlterarSenhaServicoAcessosCommand request, CancellationToken cancellationToken)
-        {
-            if (UsuarioAlterarSenhaMock.Login == request.Login && UsuarioAlterarSenhaMock.AlterarSenhaUsuarioDTOValido.SenhaAtual == request.SenhaAtual)
-            {
-                return Task.FromResult(true);
-            }
-
-            return Task.FromResult(false);
-        }
-    }
-}
