--- SME.ConectaFormacao.TesteIntegracao/Usuario/ServicosFakes/ObterPerfisUsuarioServicoAcessosPorLoginQueryHandlerFake.cs
@@ -1,20 +1,0 @@
-﻿using MediatR;
-using SME.ConectaFormacao.Aplicacao;
-using SME.ConectaFormacao.Aplicacao.DTOS;
-using SME.ConectaFormacao.TesteIntegracao.Usuario.Mocks;
-
-namespace SME.ConectaFormacao.TesteIntegracao.Usuario.ServicosFakes
-{
-    internal class ObterPerfisUsuarioServicoAcessosPorLoginQueryHandlerFake : IRequestHandler<ObterPerfisUsuarioServicoAcessosPorLoginQuery, UsuarioPerfisRetornoDTO>
-    {
-        public Task<UsuarioPerfisRetornoDTO> Handle(ObterPerfisUsuarioServicoAcessosPorLoginQuery request, CancellationToken cancellationToken)
-        {
-            if (request.Login == UsuarioRecuperarSenhaMock.LoginValido)
-            {
-                return Task.FromResult(UsuarioRecuperarSenhaMock.UsuarioPerfisRetornoDTOValido);
-            }
-
-            return Task.FromResult(new UsuarioPerfisRetornoDTO());
-        }
-    }
-}
