+++ SME.ConectaFormacao.Webapi/Controllers/PropostaController.cs
@@ -1,4 +1,5 @@
-﻿using Microsoft.AspNetCore.Mvc;
+﻿using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Mvc;
 using SME.ConectaFormacao.Aplicacao.Dtos.Proposta;
 using SME.ConectaFormacao.Aplicacao.DTOS;
 using SME.ConectaFormacao.Aplicacao.Interfaces.Proposta;
@@ -11,7 +12,7 @@ namespace SME.ConectaFormacao.Webapi.Controllers
         [ProducesResponseType(typeof(RoteiroPropostaFormativaDTO), 200)]
         [ProducesResponseType(typeof(RetornoBaseDTO), 400)]
         [ProducesResponseType(typeof(RetornoBaseDTO), 500)]
-        //[Authorize("Bearer")]
+        [Authorize("Bearer")]
         public async Task<IActionResult> ObterRoteiroPropostaFormativa([FromServices] ICasoDeUsoObterRoteiroPropostaFormativa casoDeUsoObterRoteiroPropostaFormativa)
         {
             return Ok(await casoDeUsoObterRoteiroPropostaFormativa.Executar());
